impv: refactoring modal
