+++ src/components/Modal/modal.tsx
@@ -1,46 +1,8 @@
 import React, { CSSProperties, ReactNode } from 'react';
 import { IconButton, Text } from "@chakra-ui/react"
 import { IconX } from "@tabler/icons-react"
-import { styles } from "./styles"
-
-const BACKGROUND_STYLE: CSSProperties = {
-    position: 'fixed',
-    top: 0,
-    bottom: 0,
-    left: 0,
-    right: 0,
-    backgroundColor: 'rgba(0, 0, 0, 0.7)',
-    zIndex: 1000
-}
-
-const MODAL_STYLE: CSSProperties = {
-    position: 'fixed',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    padding: '150px',
-    backgroundColor: '#FFF',
-    borderRadius: '10px'
-}
-
-const HEADER: CSSProperties = {
-    display: 'flex',
-    position: 'absolute',
-    top: '10%',
-}
-
-const TEXT_MODAL: CSSProperties = {
-    cursor: 'pointer',
-    position: 'fixed',
-    left: '10%',
-}
-
-const CLOSE_MODAL_BUTTON: CSSProperties = {
-    cursor: 'pointer',
-    borderWidth: '1px',
-    position: 'fixed',
-    right: '10%',
-}
+import { styles } from "../../pages/Associates/styles"
+import { BACKGROUND_STYLE, MODAL_STYLE, HEADER, TEXT_MODAL, CLOSE_MODAL_BUTTON } from "./styles";
 
 interface ModalProps {
     isOpen: boolean;
